fix #9 : add snake_case feature-count colnames (#10)
In a feature-counts data-frame, we updated so that (Geneid, Length), as used within featureCounts, were replaced with (feature_id, length) for greater consistency with my column-formatting conventions for data-frames.

- Renamed column-names:

    - Geneid -> `feature_id`

    - Length -> length

- Also

    - added a `.df` helper function to `helper.R` to remove duplication in the test code

    - and updated the docs

Initially was using the colnames that featureCounts outputs (Geneid, Length) alongside any sample-specific columns.

In all R work, we try to use snake_case column names. So the column names in a feature-counts dataframe should be lower-case and underscore-separated. "feature_id" is a better name than "Geneid" since we won't always be working with genes when importing from feature-counts.